feat(core): export extractSelection from /test
