+++ .changeset/extract-selection.md
@@ -1,4 +1,5 @@
 ---
+'prosekit': patch
 "@prosekit/core": patch
 ---
